+++ R/zzz.R
@@ -3,3 +3,7 @@ utils::globalVariables("snakemake")
 #' @name fastx_rename
 #' @title Rename reads in a FAST(Q/A) file
 NULL
+
+#' @name paths
+#' @title Paths for input, intermediate, and output files
+NULL